[ADD]: 💾 유저 생성 관련 함수 싱글톤으로 이동
